Move python compatibility functions to compat module
