+++ mecode/compat.py
@@ -1,0 +1,22 @@
+try:
+    isinstance('', basestring)
+
+    def is_str(s):
+        return isinstance(s, basestring)
+
+    def encode2To3(s):
+        return s
+
+    def decode2To3(s):
+        return s
+
+except NameError:
+
+    def is_str(s):
+        return isinstance(s, str)
+
+    def encode2To3(s):
+        return bytes(s, 'UTF-8')
+
+    def decode2To3(s):
+        return s.decode('UTF-8')
